move code from header to cpp
